requirements.txt
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.4
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.92.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwtdown-fastapi==0.5.0
MarkupSafe==2.1.3
orjson==3.9.2
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
psycopg>=3.1.2
psycopg-binary>=3.1.2
psycopg2-binary==2.9.1
psycopg-pool>=3.1.7
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.11
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.31
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.17.6
uvloop==0.17.0
watchgod==0.8.2
websockets==11.0.3
